main added

+++ managers.py
@@ -2,6 +2,7 @@ import cv2
 import numpy
 import time
 
+
 class CaptureManager(object):
 
     def __init__(self, capture, previewWindowManager = None, shouldMirrorPreview = False):
@@ -22,7 +23,7 @@ class CaptureManager(object):
         self._videoWriter = None
 
         self._startTime = None
-        self._framesElapsed = long(0)
+        self._framesElapsed = 0
         self._fpsEstimate = None
     
     @property
